requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
requests
python-dateutil
streamlit
psycopg2-binary
sqlalchemy
fastapi
uvicorn
